requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
Scrapy==2.12.0
google-generativeai